+++ validator/client/BUILD.bazel
@@ -39,6 +39,7 @@ go_library(
         "//validator/keymanager:go_default_library",
         "//validator/keymanager/imported:go_default_library",
         "//validator/slashing-protection/iface:go_default_library",
+        "//validator/slashing-protection/local/attesting-history:go_default_library",
         "@com_github_dgraph_io_ristretto//:go_default_library",
         "@com_github_gogo_protobuf//proto:go_default_library",
         "@com_github_gogo_protobuf//types:go_default_library",
