+++ validator/slashing-protection/local/attesting-history/BUILD.bazel
@@ -1,0 +1,27 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_test")
+load("@prysm//tools/go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["check_slashable_attestation.go"],
+    importpath = "github.com/prysmaticlabs/prysm/validator/slashing-protection/local/attesting-history",
+    visibility = ["//validator:__subpackages__"],
+    deps = [
+        "//shared/params:go_default_library",
+        "//validator/db/kv:go_default_library",
+        "@com_github_pkg_errors//:go_default_library",
+        "@com_github_sirupsen_logrus//:go_default_library",
+    ],
+)
+
+go_test(
+    name = "go_default_test",
+    srcs = ["check_slashable_attestation_test.go"],
+    embed = [":go_default_library"],
+    deps = [
+        "//shared/bytesutil:go_default_library",
+        "//shared/params:go_default_library",
+        "//shared/testutil/require:go_default_library",
+        "//validator/db/kv:go_default_library",
+    ],
+)
