+++ validator/slashing-protection/local/standard-protection-format/BUILD.bazel
@@ -14,9 +14,9 @@ go_library(
     deps = [
         "//shared/bytesutil:go_default_library",
         "//shared/hashutil:go_default_library",
-        "//validator/client:go_default_library",
         "//validator/db:go_default_library",
         "//validator/db/kv:go_default_library",
+        "//validator/slashing-protection/local/attesting-history:go_default_library",
         "@com_github_k0kubun_go_ansi//:go_default_library",
         "@com_github_pkg_errors//:go_default_library",
         "@com_github_schollz_progressbar_v3//:go_default_library",
